youtube/content_planner/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-api-python-client==2.116.0
python-dotenv==1.0.0
httpx==0.26.0
youtube-transcript-api==1.2.4
google-generativeai>=0.8.0
